ui_tests/requirements.txt
pytest-splunk-addon
selenium
webdriver_manager
pyyaml